refactor: telemetry implementation
